+++ src/engine/camera/camera.h
@@ -19,7 +19,7 @@ public:
     virtual SDL_Rect getViewpoit() = 0;
     virtual Transform::TransformRef getTransform() = 0;
     virtual void setTransform(Transform::TransformRef) = 0;
-    virtual Vector2<double> screenToWorldPosition(Vector2<int> const&) = 0;
+    virtual Vector2<float> screenToWorldPosition(Vector2<int> const&) = 0;
 };
 
 class Camera : public ICamera
@@ -42,7 +42,7 @@ public:
     virtual SDL_Rect getViewpoit() { return viewport; }
     virtual Transform::TransformRef getTransform() { return transform; }
     virtual void setTransform(Transform::TransformRef value) { transform = value; }
-    virtual Vector2<double> screenToWorldPosition(Vector2<int> const&);
+    virtual Vector2<float> screenToWorldPosition(Vector2<int> const&);
 
 private:
     void updateViewport();
